requirements.txt
Flask==3.0.0
PyJWT==1.7.1
flask-cors
gunicorn
cryptography
gunicorn